♻️ Standardize luck and rarity on u8 instead of u16
Fruit._rarity, Member._luck, and Community._luck are now u8 (0–255).
Rarity values drop the ×257 scaling factor; luck getters/setters divide
by u8::MAX. EventPayload and StateMutation luck fields follow suit.
Spec and test tolerances updated to match u8 precision.

+++ command_line_service/src/repl.rs
@@ -305,7 +305,7 @@ fn cmd_luck(
             return;
         }
     };
-    let luck = (luck_f64 * u16::MAX as f64).round() as u16;
+    let luck = (luck_f64 * u8::MAX as f64).round() as u8;
 
     let community = fetch(store, id);
